Renaming scala packages to declarativewidgets (c) Copyright IBM Corp. 2016
