+++ kernel-scala/src/main/scala/declarativewidgets/exceptions/WidgetNotAvailableException.scala
@@ -3,7 +3,7 @@
  * Distributed under the terms of the Modified BSD License.
  */
 
-package urth.widgets.exceptions
+package declarativewidgets.exceptions
 
 case class WidgetNotAvailableException(klass: String)
   extends RuntimeException(s"Scala widget class for $klass not found!")
