+++ kernel-scala/src/test/scala/declarativewidgets/WidgetDataFrameSpec.scala
@@ -3,7 +3,7 @@
  * Distributed under the terms of the Modified BSD License.
  */
 
-package urth.widgets
+package declarativewidgets
 
 import org.apache.spark.SharedSparkContext
 import org.apache.spark.sql.{SQLContext, DataFrame}
